chore(tools): separate out metabase reading functions (#4825)
